Add support for ltex-ls client
ltex-ls uses the `data` field in requests and
needs to have complete support for it. It turns
out that ltex-ls also requires a fix on the
client side because they publish the HTTP request
with the wrong Content-Type, so axum does not
even forwards the request to check. This is fixed
by:

https://github.com/Javier-varez/ltex-ls/commit/bb116c7dbac199b10d5df25bdadd878aa8dc72ca